add copy and fc in colsole

+++ VorobieiYevhen/src/week2/concole/MyFileHelper.java
@@ -274,39 +274,67 @@ public class MyFileHelper implements FileHelper {
 
     @Override
     public boolean copy() {
-        String fileFrom = null;
-        String fileTo = null;
 
-        FileOutputStream out = null;
-        FileInputStream in = null;
+        String fileTo;
 
+        String fileFrom = getFilePath();
         try {
-            in = new FileInputStream(fileFrom);
-            out = new FileOutputStream(fileTo);
-            return  IOUtils.copy(in, out);
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } finally {
+            if (!new File(fileFrom).isFile()) {
+                System.out.println("Can't read, this is not a file!");
+            } else  {
+
+                String[] fileToArray = fileFrom.split("\\.");
+                fileTo = fileToArray[0] + "(copy)." + fileToArray[1];
+                FileOutputStream out = null;
+                FileInputStream in = null;
+
                 try {
-                    if (in != null){
-                        in.close();
-                    }
-                } catch (IOException e) {
+                    in = new FileInputStream(fileFrom);
+                    out = new FileOutputStream(fileTo);
+                    OutputStream file = IOUtils.copy(in, out);
+                } catch (FileNotFoundException e) {
                     e.printStackTrace();
+                } finally {
+                    IOUtils.closeIn(in);
+                    IOUtils.closeOut(out);
                 }
-            try {
-                if (out != null){
-                    out.flush();
-                    out.close();
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
             }
-
-
+        } catch (NullPointerException e) {
+            System.out.println("Can't read this file! (Unreadable format)");
         }
 
-        return false;
+        return file != null ? true :false;
+    }
+
+    @Override
+    public boolean fc() {
+        FileInputStream fileOne = null;
+        FileInputStream fileTwo = null;
+        try {
+            String filePathOne = getFilePath();
+            if (!new File(filePathOne).isFile()) {
+                System.out.println("It's not a file");
+                return false;
+            }
+            String filePathTwo = getFilePath();
+            if (!new File(filePathTwo).isFile()) {
+                System.out.println("It's not a file");
+                return false;
+            }
+            fileOne = new FileInputStream(filePathOne);
+            fileTwo = new FileInputStream(filePathTwo);
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+            System.out.println("File Not Found!");
+            return false;
+        }catch (NullPointerException e) {
+            System.out.println("Can't read this file!");
+            return false;
+        }
+        String fileOneString = IOUtils.fileInBytes(fileOne);
+        String fileTwoString = IOUtils.fileInBytes(fileTwo);
+
+        return fileOneString.equals(fileTwoString) ? true : false;
     }
 
     private String getFilePath() {
